Adds sawyer_robot_hw_sim to customize Gazebo
This adds a new plugin to inheret from
gazebo_ros_control::default_robot_hw_sim in order
to customize some of the feel for the simulator.
In particular, this dampens the joints to prevent
movement when the robot is disabled, and returns
the damping to normal when it is enabled again.