ArmControllerInterface cleanup and bugfixes
Cleans up some of the logic and code structure for the
ArmControllerInterface. Turns out the controllers work much
better now, because a logic bug was fixed which caused the
controllers to start and stop on every new command coming in.

Now the controllers only start and stop when the command mode
is changed (same functionality as the real robot). Sawyer now
responds much better to position commands.