Add SharedJointInterface; Gravity Compensation controller
Introduce the SumJointHandle and SharedJointInterface, which allow for
multiple controllers to command a single joint by summing the commanded
values from each controller into a final output to the sim-joint.

The SharedJointInterface takes the place of the EffortJointInterface,
from which controllers request JointHandles. The JointHandles behave
exactly the same, but multiple unique handles can be commanded at the
same time by a controller requesting a JointHandle with an additional
'sub-type' parameter.

This allows the introduction of the GravityCompensation Controller,
which adds a parallel component to offset gravity to the commanded effort
for each joint.

Major pieces:

* Add SumInterface
* Add SharedJointInterface
  - replacement for EffortJointInterface in Sawyer Controllers
* Update commandSum in SimRobotHW write loop