Add copies of upstream effort_controllers to take SharedHWInterface
The normal ros_controllers/effort_controllers used by the
SawyerControllers to control each joint do not allow a subclass of the
EffortJointInterface to be passed in as a replacement. Thus, we need to
(temporarily) copy in the upstream effort_controllers so that we can
change them to take the SharedJointInterface.

We should remove these copies once we can upstream a change to make the
init() and constructor virtual (or some similar modification) so that
interfaces that inherit from EffortJointInterface can be used instead.