Updates Camera Sim IO controller config pubs
The camera example and intera_interface.Cameras() class
broke some of the assumptions used for the camera_sim_io
node, particularly with regards to not utilizing
the config. This adds in config messages directly from
an SDK robot. It also makes sure all messages are considered
"valid" by forcing the minimum timestamp to be 1 second.
A zero timestamp is considered invalid by the IO Framework.