Adds joint_limits pub to Arm Kinematics Interface
To comply with the new JointLimits class in the
intera_interface Python API, this adds a new publisher
to the arm_kinematics_interface. That publisher
sends out all of the joint limits from the URDF and
acceleration config files over the <namespace>/joint_limits
topic.

This new JointLimits object allowed an optimization of
the kinematic chain creation for SNS_IK, as we don't
need to look for the joint limits for every joint in the
chain anymore. Rather, we can look up each joint in
the JointLimits message (which is only constructed once),
and use them directly.