Adds disabling of Gravity Compensation Torques
This creates a newly subscribed topic:
/robot/limb/right/suppress_gravity_compensation
which takes a std_msgs/Empty message. If this topic
is published to at >= 5.0 Hz, all gravity compensation
torque additions are disabled.

The user can publish directly from the commandline:
$ rostopic pub /robot/limb/right/suppress_gravity_compensation
std_msgs/Empty "{}" -r 5

This leads to poorer tracking in Position, Trajectory,
and Velocity modes as each mode's PID controller is tuned
for gravity torques to be present, and has to work harder
to hold up the arm.
In Torque control mode, this causes the arm to drop down
completely unless the user is providing torques to counter
gravity.